Improved double jump
I've improved the jump. before you had to jump at least on a platform to be allowed to jump again. With the new change you can now jump 1 time at least when you fall from a platform instead of having to jump to be allowed to jump mid-air. Also gave ground component to walls, so you can now jump when standing on top of walls